Move new doc-lb widget so it doesn't make top part of window taller
It does take away horizontal space from other things, because now it has been
moved between the Vars list and the doc text area.

It also doesn't behave the way I want when you resize the window -- it
often leaves a gap of nothing to the left of the doc text area.  Need
to figure out how to make it pack so that there is no such gap.