New menu items "Search Source Also" and "Search Examples Also"
Issues that would be good to fix:

"Search Examples Also" is currently enabled, and can be toggled off
and on, even when in ClojureDocs on-line mode.  However the examples
are only searched when in ClojureDocs off-line mode.  One way to make
that more obvious to the user is to disable the "Search Examples Also"
menu item when in ClojureDocs on-line mode.

When I have "Search Examples Also" selected, and have a string that
I've typed in to search through the examples for, then go to
ClojureDocs on-line mode, then switch back to ClojureDocs off-line
mode, the UI hangs.  I don't know why yet.